backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
mysql-connector-python==9.0.0
python-dotenv==1.0.1
email-validator==2.2.0
